uqmi: move connection logic to qmi.script
Move the portion of the netifd handler that deals with setting up IP
connectivity for QMI interfaces to a script that is started using
proto_run_command.

Before this patch, addresses and routes were configured on the main
interface with dhcp/dhcpv6 = 0. With this patch, there's always a _{4,6}
child interface (like there was when DHCP was used).

The script monitors connection state for both address families using the
modem and restarts the corresponding child interface.

This also allows selectively bringing up IPv4/IPv6 without restarting the
main interface (which would include restarting the registration process).

The portion of the protocol handler that deals with unlocking the SIM
and registering with the network remains in the handler for no better
reason than the lack of a good idea where to put it.

Signed-off-by: Leon M. Busch-George <leon@georgemail.eu>